+++ timermode.h
@@ -7,7 +7,7 @@ class TimerMode // super class for all the modes for the timer
 protected:  // protected so that the subclasses can change them freely
             // but they are encapsulated from the rest of the program
 
-    float timeEarned;      // total time earned
+    double timeEarned;      // total time earned
     int timeRunning;        // time in seconds since the hourglass was last flipped
     bool timerDirection;    // 0 = negative and 1 = positive
     int difficulty;         // difficulty is used in the calcTimeEarned
@@ -17,12 +17,12 @@ public:
     ~TimerMode();
     void timerFlip();           // inverses the timers direction
     bool getTimerDirection() const;
-    float getTimeEarned() const;
+    double getTimeEarned() const;
 
     void calcTimeEarned(); // For some reason this being virtual breaks the static_casting
     int getTimeRunning() const;
     void updateTimeRunning();   // increments the time running
-    void setTimeEarned(float newTimeEarned);    // used for deducting timeEarned when purchasing mode
+    void setTimeEarned(double newTimeEarned);    // used for deducting timeEarned when purchasing mode
     int getDifficulty() const;
     void setDifficulty(int newDifficulty);
     void resetTimeRunning();
